requirements.txt
archon>=0.1.0
langchain>=0.1.0
pydantic>=2.0.0
chromadb>=0.4.0
supabase>=2.0.0
duckduckgo-search>=4.1.1
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.36.0 